#11 moved all models to Realm; created extensions for each model
