+++ EAO/PFAudio.swift
@@ -23,10 +23,10 @@ final class PFAudio: Object, Mappable{
         static let coordinate = "coordinate"
     }
     
-    @objc dynamic var id            : String = UUID().uuidString
+    @objc dynamic var id            : String = "\(UUID().uuidString).mp4a"
     @objc dynamic var observationId : String?
     @objc dynamic var inspectionId : String?
-    @objc dynamic var index: NSNumber? = 0
+    @objc dynamic var index: Int = 0
     @objc dynamic var notes: String?
     @objc dynamic var title: String?
     @objc dynamic var url: String?
@@ -69,20 +69,59 @@ extension PFAudio {
     }
     
     @objc func get() -> Data?{
-        //        guard let id = id else{
-        //            return nil
-        //        }
-        //        let url = URL(fileURLWithPath: FileManager.directory.absoluteString).appendingPathComponent(id, isDirectory: true)
-        //        return try? Data(contentsOf: url)
-        return nil
+        let url = URL(fileURLWithPath: FileManager.directory.absoluteString).appendingPathComponent(id, isDirectory: true)
+        return try? Data(contentsOf: url)
+    }
+    
+    @objc func getURL() -> URL?{
+        let url = URL(fileURLWithPath: FileManager.directory.absoluteString).appendingPathComponent(id, isDirectory: true)
+        return url
+    }
+}
+
+/*
+final class PFAudio: PFObject, PFSubclassing{
+    
+    @NSManaged var id            : String?
+    @NSManaged var observationId : String?
+    @NSManaged var inspectionId : String?
+    @NSManaged var coordinate    : PFGeoPoint?
+    @NSManaged var index: NSNumber?
+    @NSManaged var notes: String?
+    @NSManaged var title: String?
+    @NSManaged var url: URL?
+    @NSManaged var file : PFFile?
+    
+    static func parseClassName() -> String {
+        return "audio"
+    }
+    
+    @objc static func load(for observationId: String, result: @escaping (_ audios: [PFAudio]?)->Void){
+        guard let query = PFAudio.query() else{
+            result(nil)
+            return
+        }
+        query.fromLocalDatastore()
+        query.whereKey("observationId", equalTo: observationId)
+        query.findObjectsInBackground(block: { (audios, error) in
+            result(audios as? [PFAudio])
+        })
+    }
+    
+    @objc func get() -> Data?{
+        guard let id = id else{
+            return nil
+        }
+        let url = URL(fileURLWithPath: FileManager.directory.absoluteString).appendingPathComponent(id, isDirectory: true)
+        return try? Data(contentsOf: url)
     }
     
     @objc func getURL() -> Data?{
-        return nil
-        //        guard let id = id else{
-        //            return nil
-        //        }
-        //        let url = URL(fileURLWithPath: FileManager.directory.absoluteString).appendingPathComponent(id, isDirectory: true)
-        //        return try? Data(contentsOf: url)
+        guard let id = id else{
+            return nil
+        }
+        let url = URL(fileURLWithPath: FileManager.directory.absoluteString).appendingPathComponent(id, isDirectory: true)
+        return try? Data(contentsOf: url)
     }
 }
+ */
