+++ EAO/PFVideo.swift
@@ -23,10 +23,10 @@ final class PFVideo: Object, Mappable{
         static let coordinate = "coordinate"
     }
     
-    @objc dynamic var id            : String = UUID().uuidString
+    @objc dynamic var id            : String = "\(UUID().uuidString).mp4"
     @objc dynamic var observationId : String?
     @objc dynamic var inspectionId : String?
-    @objc dynamic var index: NSNumber? = 0
+    @objc dynamic var index: Int = 0
     @objc dynamic var notes: String?
     @objc dynamic var title: String?
     @objc dynamic var url: String?
@@ -66,23 +66,61 @@ final class PFVideo: Object, Mappable{
     }
     
     @objc func get() -> Data?{
-        //        guard let id = id else{
-        //            return nil
-        //        }
-        //        let url = URL(fileURLWithPath: FileManager.directory.absoluteString).appendingPathComponent(id, isDirectory: true)
-        //        print(url)
-        //        print("\(try? Data(contentsOf: url).count)")
-        //        return try? Data(contentsOf: url)
-        return nil
+        let url = URL(fileURLWithPath: FileManager.directory.absoluteString).appendingPathComponent(id, isDirectory: true)
+        return try? Data(contentsOf: url)
     }
     
     @objc func getURL() -> URL?{
-        //        guard let id = id else{
-        //            return nil
-        //        }
-        //        let url = URL(fileURLWithPath: FileManager.directory.absoluteString).appendingPathComponent(id, isDirectory: true)
-        //        return try? url
-        return nil
+        let url = URL(fileURLWithPath: FileManager.directory.absoluteString).appendingPathComponent(id, isDirectory: true)
+        return url
     }
 }
 
+/*
+final class PFVideo: PFObject, PFSubclassing{
+    
+    @NSManaged var id            : String?
+    @NSManaged var observationId : String?
+    @NSManaged var inspectionId : String?
+    @NSManaged var coordinate    : PFGeoPoint?
+    @NSManaged var index: NSNumber?
+    @NSManaged var notes: String?
+    @NSManaged var title: String?
+    @NSManaged var url: URL?
+    @NSManaged var file : PFFile?
+    
+    static func parseClassName() -> String {
+        return "video"
+    }
+    
+    @objc static func load(for observationId: String, result: @escaping (_ videos: [PFVideo]?)->Void){
+        guard let query = PFVideo.query() else{
+            result(nil)
+            return
+        }
+        query.fromLocalDatastore()
+        query.whereKey("observationId", equalTo: observationId)
+        query.findObjectsInBackground(block: { (videos, error) in
+            result(videos as? [PFVideo])
+        })
+    }
+    
+    @objc func get() -> Data?{
+        guard let id = id else{
+            return nil
+        }
+        let url = URL(fileURLWithPath: FileManager.directory.absoluteString).appendingPathComponent(id, isDirectory: true)
+        print(url)
+        print("\(try? Data(contentsOf: url).count)")
+        return try? Data(contentsOf: url)
+    }
+    
+    @objc func getURL() -> URL?{
+        guard let id = id else{
+            return nil
+        }
+        let url = URL(fileURLWithPath: FileManager.directory.absoluteString).appendingPathComponent(id, isDirectory: true)
+        return try? url
+    }
+}
+*/
